Function for creating objects. DerivedUpdate now takes a double instead of a float

+++ PA6/include/object.h
@@ -9,14 +9,14 @@
 class Object : public BaseObject{
 public:
     Object(BaseObject *parent_, std::string objectPath);
-    void DerivedUpdate(float dt);
+    void DerivedUpdate(double dt);
 
     virtual void MouseDown(SDL_Event event);
     virtual void MouseUp(SDL_Event event);
     virtual void KeyDown(SDL_Event event);
     virtual void KeyUp(SDL_Event event);
     virtual void MouseWheel(SDL_Event event);
-    
+
 private:
     float rotateRate;
